Creating routes via vue-router
