+++ package.json
@@ -23,7 +23,7 @@
     "ts-node": "^10.9.2",
     "tsx": "^4.20.3",
     "vue": "^3.5.17",
-    "vue-router": "^4.5.1"
+    "vue-router": "4"
   },
   "devDependencies": {
     "@types/compression": "^1.8.1",
