+++ src/layouts/DefaultLayout.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts"></script>
+<template>
+  <main
+    class="bg-emerald-600 text-foreground min-h-[100dvh] overflow-x-hidden w-[100vw] max-w-[100vw] px-2"
+  >
+    <RouterView />
+  </main>
+</template>
+
+<style scoped></style>
